PI-452 refactored permissions endpoints into viewsets instead of views
